% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{button}
\alias{button}
\title{buttons with status}
\usage{
button(id, label, icon, status)
}
\arguments{
\item{id}{A character string of the shiny id.}

\item{label}{A characer string of the label.}

\item{icon}{A character string of the icon.}

\item{status}{A character string of the status.}
}
\value{
a shiny tag.
}
\description{
buttons with status
}
