% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtilities.R
\name{createImage}
\alias{createImage}
\title{createImage}
\usage{
createImage(src, id, width = "20", height = "20", width.mouseHover = "22",
  height.mouseHover = "22", class = NULL)
}
\arguments{
\item{src}{Image source.}

\item{id}{Unique html id.}

\item{width}{Regular width.}

\item{height}{Regular height.}

\item{width.mouseHover}{Width when hovering over image.}

\item{height.mouseHover}{Height when hovering over image.}
}
