% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css_body}
\alias{css_body}
\title{Body css}
\usage{
css_body(
  font_family = "Myriad-Pro, Calibri, Arial, 'sans serif'",
  font_weight = "100",
  font_size_body = "10px",
  font_size_label = "13px",
  text_color = "#494949"
)
}
\arguments{
\item{font_family}{A character string of body font family.}

\item{font_weight}{A character string of body font weight.}

\item{font_size_body}{A character string of body font size.}

\item{font_size_label}{A character string of label font size.}

\item{text_color}{A character string of body text color.}
}
\value{
a shiny tag.
}
\description{
Body css
}
