% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css_leaflet_full}
\alias{css_leaflet_full}
\title{Leaflet full screen}
\usage{
css_leaflet_full(x = 100)
}
\arguments{
\item{x}{A number of the percentage vertical height}
}
\value{
a shiny tag.
}
\description{
Leaflet full screen
}
