% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css_navbar}
\alias{css_navbar}
\title{Navbar css}
\usage{
css_navbar(
  text_color = "#E0E0E0",
  text_selected_color = "#37CBA2",
  text_size = "15px",
  background_color = "#010101",
  float = "right"
)
}
\arguments{
\item{text_color}{A character string of navbar text color.}

\item{text_selected_color}{A character string of selected navbar text color.}

\item{text_size}{A character string of navbar text size.}

\item{background_color}{A character string of navbar background color.}

\item{float}{A character string indicating position of tabs in navbar ('left', 'right' or 'center')}
}
\value{
a shiny tag.
}
\description{
Navbar css
}
