% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css_slider}
\alias{css_slider}
\title{Slider css}
\usage{
css_slider(slider_color = "#494949", background_color = "#101010")
}
\arguments{
\item{slider_color}{A character string of slider color.}

\item{background_color}{A character string of slider background color.}
}
\value{
a shiny tag.
}
\description{
Slider css
}
