% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{fix_text}
\alias{fix_text}
\title{text for signalling something should be fixed}
\usage{
fix_text(x, font_size = "13px")
}
\arguments{
\item{x}{A character string .}

\item{font_size}{A character string of the font size.}
}
\value{
a shiny tag.
}
\description{
text for signalling something should be fixed
}
