% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{help_size}
\alias{help_size}
\title{help tag and font size}
\usage{
help_size(x, font_size = "13px")
}
\arguments{
\item{x}{A character string .}

\item{font_size}{A character string of the font size.}
}
\value{
a shiny tag.
}
\description{
help tag and font size
}
