% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{info}
\alias{info}
\title{Info text appear when icon clicked}
\usage{
info(
  id,
  label,
  help,
  label_size = "15px",
  help_size = "12px",
  icon_size = "15px",
  ns = NS(NULL),
  ...
)
}
\arguments{
\item{id}{A character string of the shiny id.}

\item{label}{A character string of the package name.}

\item{help}{A taglist containing ui elements for the upper body.}

\item{label_size}{A character string of the label font size.}

\item{help_size}{A character string of the help font size.}

\item{icon_size}{A character string of the icon size.}

\item{ns}{A function of the namespace.}

\item{...}{arguments passed to actionLink}
}
\value{
a shiny taglist
}
\description{
Info text appear when icon clicked
}
