% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtilities.R
\name{insertIcon}
\alias{insertIcon}
\title{insertIcon}
\usage{
insertIcon(src, id, width = "20", height = "20", width.mouseHover = "22",
  height.mouseHover = "22", class = NULL, name = NULL)
}
\arguments{
\item{src}{Image source.}

\item{id}{Unique html id.}

\item{width}{Regular width.}

\item{height}{Regular height.}

\item{width.mouseHover}{Width when hovering over image.}

\item{height.mouseHover}{Height when hovering over image.}

\item{class}{Class of the newly created image.}
}
