% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtilities.R
\name{insertTab}
\alias{insertTab}
\title{insertTab}
\usage{
insertTab(id, title, renderExpression, placeholder = "#tabsetPlaceholder")
}
\arguments{
\item{id}{Unique html id of element to be inserted.}

\item{title}{Displayed tab title.}

\item{renderExpression}{An expression to be displayed in a div on the newly created tab.}
}
