% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{label_size}
\alias{label_size}
\title{label tag and font size}
\usage{
label_size(x, font_size = "13px")
}
\arguments{
\item{x}{A character string .}

\item{font_size}{A character string of the font size.}
}
\value{
a shiny tag.
}
\description{
label tag and font size
}
