% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{mod_about_ui}
\alias{mod_about_ui}
\alias{mod_about_server}
\title{mod_about_ui and mod_about_server}
\usage{
mod_about_ui(id, package = "shinyutils",
  header = poisson_header(package),
  citation = poisson_citation(package), footer = poisson_footer(), ...)

mod_about_server(input, output, session)
}
\arguments{
\item{id}{A character string of the shiny id.}

\item{package}{A character string of the package name.}

\item{header}{A taglist containing ui elements for the upper body.}

\item{citation}{A taglist containing ui elements for the citation.}

\item{footer}{A taglist containing ui elements for the lower body.}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to generate About tab.
}
\keyword{internal}
