% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{p_size}
\alias{p_size}
\title{p tag and font size}
\usage{
p_size(x, font_size = "13px")
}
\arguments{
\item{x}{A character string .}

\item{font_size}{A character string of the font size.}
}
\value{
a shiny tag.
}
\description{
p tag and font size
}
