% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{poisson_citation}
\alias{poisson_citation}
\title{Poisson package citation}
\usage{
poisson_citation(package = "shinyutils")
}
\arguments{
\item{package}{A character string of the package name.}
}
\value{
a shiny taglist.
}
\description{
Generate UI for upper part of About tab body
}
