% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyUtilities.R
\name{renderVersion}
\alias{renderVersion}
\title{renderVersion}
\usage{
renderVersion(url, version, deploymentDate)
}
\arguments{
\item{url}{Url to source code repository}

\item{version}{Version vector containing four integers}

\item{deploymentDate}{Date formatted as String}
}
\description{
renderVersion
}

