% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{select_input_x}
\alias{select_input_x}
\title{Select input with x}
\usage{
select_input_x(
  id,
  label = "Select sites:",
  choices,
  selected = choices[1],
  ...
)
}
\arguments{
\item{id}{A character string of the shiny id.}

\item{label}{A characer string of the label.}

\item{choices}{A vector of characer strings of choices.}

\item{selected}{A characer string of the initial selected choice.}

\item{...}{arguments passed to selectizeInput}
}
\value{
a shiny tag.
}
\description{
Select input with x
}
