% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_bot}
\alias{slack_bot}
\title{Pretty slack bot message}
\usage{
slack_bot(..., channel, username, incoming_webhook_url)
}
\arguments{
\item{...}{An expression.}

\item{channel}{A character string of the channel name.}

\item{username}{A character string of the username.}

\item{incoming_webhook_url}{A character string of the incoming webhook url.}
}
\value{
TRUE
}
\description{
Pretty slack bot message
}
