% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{value_box}
\alias{value_box}
\title{Value box font size}
\usage{
value_box(subtitle, value, color = "green", font_size = "70\%", ...)
}
\arguments{
\item{subtitle}{A character string of the subtitle.}

\item{value}{A character string of the value.}

\item{color}{A character string of the color.}

\item{font_size}{A character string of the font size.}

\item{...}{arguments passed to valueBox}
}
\value{
a shiny tag.
}
\description{
Value box font size
}
